requirements.txt
aiosqlite
discord.py
python-dotenv
sqlalchemy[asyncio]
